requirements.txt
requests
python-dotenv
configparser